Try typecasting size_t to int

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -33,7 +33,7 @@ int interpolation_search(int *array, size_t size, int value)
 		else if (array[pos] > value)
 			high = pos - 1;
 		else
-			return (pos);
+			return ((int)pos);
 	}
 	return (-1);
 }
